core: tee_mm.c: pull architecture-independent code from arch folder
When porting OP-TEE OS to another architecture we end up with a huge amount
of code being duplicated. The purpose of this commit is to check the
feasibility of re-factoring the project tree to sort out
architecture-independent code to ease the portability and the maintenance
for other archs. I started with moving tee_mm.c to core/mm folder.

Signed-off-by: Marouene Boubakri <marouene.boubakri@nxp.com>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>